requirements.txt
fastapi
uvicorn
python-dotenv
requests
sqlalchemy
google-api-python-client
